fix: カスタムフック名の変更とApp.jsxに記載したsetFnameなどをカスタムフックに移動
